fix: move config initialization to separate module to resolve circular imports
