+++ openhands/server/file_config.py
@@ -3,7 +3,7 @@ import re
 
 from openhands.core.config import AppConfig
 from openhands.core.logger import openhands_logger as logger
-from openhands.server.shared import config as shared_config
+from openhands.server.config_init import config as shared_config
 
 FILES_TO_IGNORE = ['.git/', '.DS_Store', 'node_modules/', '__pycache__/', 'lost+found/']
