+++ openhands/server/services/github_service.py
@@ -6,7 +6,7 @@ from fastapi import Request
 from openhands.server.auth import get_github_token
 from openhands.server.data_models.gh_types import GitHubRepository, GitHubUser
 from openhands.server.settings_store import SettingsStoreImpl
-from openhands.server.shared import config, server_config
+from openhands.server.config_init import config, server_config
 from openhands.server.types import AppMode, GhAuthenticationError, GHUnknownException
