+++ openhands/server/shared.py
@@ -1,22 +1,12 @@
 import os
-
 import socketio
-from dotenv import load_dotenv
 
-from openhands.core.config import load_app_config
-from openhands.server.config.server_config import load_server_config
+from openhands.server.config_init import config, server_config, file_store
 from openhands.server.conversation_manager.conversation_manager import (
     ConversationManager,
 )
-from openhands.storage import get_file_store
 from openhands.storage.conversation.conversation_store import ConversationStore
 from openhands.utils.import_utils import get_impl
-
-load_dotenv()
-
-config = load_app_config()
-server_config = load_server_config()
-file_store = get_file_store(config.file_store, config.file_store_path)
 
 client_manager = None
 redis_host = os.environ.get('REDIS_HOST')
